Add files via upload
This is a copy of a CSV file of all companies listed in various stock market exchanges found on the Nasdaq website: 

https://www.nasdaq.com/market-activity/stocks/screener

The file can be downloaded directly from the website and includes various metrics which seem to be updated daily (such as Last Sale, Net Change, etc).  We are using this file primarily in order to populate a MySQL database to include a snapshot of all available symbols, including any interesting data which we can gleam during the snapshot including market cap, price (and therefore calculated float), IPO Year, Sector, Industry, etc.  This saves us having to manually populate a stock symbol database table.